12.9 修改了取消下载的逻辑，还是单线程逻辑 Signed-off-by: chsl926 <350306656@qq.com>

+++ Assets/Scripts/Lobby.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using UnityEngine.UI;
 using System.IO;
 using YZL.Compress.UPK;
+using System;
 
 public class Lobby : MonoBehaviour
 {
@@ -17,14 +18,25 @@ public class Lobby : MonoBehaviour
     private bool loadluaend = false;
     public string TestSceneName;
     private List<string> delReList=new List<string>();
+    private Button ClickButton;
+    private bool isMaskAni = false;
+    private bool isCanelend = false;
+    private int  maskAniPos = -1;
+    private static float releaseReVal = 0;
+    private static  List<string> cancelList = new List<string>();
     private enum State
     {
         NOMAL,
-        EDIT
+        EDIT,
+        DOWNLOAD,
+        UNDOWNlOAD,
+        UNZIP,
+        CANCELUNZIP,
     }
     private State Curstate=State.NOMAL;
     void Start()
     {
+        SDK.CancelBack = OnCancelLoad;
         ShowGui.info = "";
         SDK.ClearSDK();
         if (SDK.sdk == null)
@@ -176,12 +188,16 @@ public class Lobby : MonoBehaviour
             Init();
             isinit = true;
         }
-        if (loadreend == true && loadluaend == true)
-        {
-            GoMainScene();
-            loadreend = false;
-            loadluaend = false;
-        }
+        //if (loadreend == true && loadluaend == true)
+        //{
+        //    GoMainScene();
+        //    loadreend = false;
+        //    loadluaend = false;
+        //}
+        //播放按钮的mask填充动画
+        MaskAni();
+        //清除取消的资源
+        ClearCanelRe();
     }
     //初始化
     void Init()
@@ -192,7 +208,7 @@ public class Lobby : MonoBehaviour
         ShowGui.info = "开始下载： " + SimpleFramework.Util.DataPath + "\n";
         if (File.Exists(SimpleFramework.Util.DataPath + "SceneData.xml"))
             File.Delete(SimpleFramework.Util.DataPath + "SceneData.xml");
-        SDK.DoDownLoadText("SceneData.xml", SimpleFramework.Util.DataPath, "", OnUpSceneList);
+        SDK.DoDownLoadText("SceneData.xml", SimpleFramework.AppConst.AppName, "", OnUpSceneList);
     }
     //检查游戏场景列表
     void OnUpSceneList(string txt)
@@ -203,11 +219,11 @@ public class Lobby : MonoBehaviour
         for (int i = 0; i < info.Count; i++)
         {
             string time = info[i].time;
-            SDK.DownLoadImage("icon/" + time + ".png", SimpleFramework.Util.DataPath + "Image", "time");
+            SDK.DownLoadImage("icon/" + time + ".png", SimpleFramework.AppConst.AppName + "/Image", "time");
             for (int j = 0; j < info[i].games.Length; j++)
             {
                 string game = info[i].games[j];
-                SDK.DownLoadImage("icon/" + game + ".png", SimpleFramework.Util.DataPath + "Image", "game");
+                SDK.DownLoadImage("icon/" + game + ".png", SimpleFramework.AppConst.AppName + "/Image", "game");
             }
            
         }
@@ -243,15 +259,50 @@ public class Lobby : MonoBehaviour
             {
                 newButton.transform.FindChild("Mask").gameObject.SetActive(false);
             }
+
             //点击按钮
             newButton.GetComponent<DownTime>().ClickEvent=(end)=>
              {
-                 //正在初始化资源
+                 //正在初始化资源返回
                  if (!File.Exists(SimpleFramework.Util.DataPath + "files.txt") || !Directory.Exists(SimpleFramework.Util.DataPath + "lua"))
                      return;
-                 //正在编辑资源
-                 if (Curstate == State.EDIT) return;
-                 CheckResoucre(name);
+                 //清除已经取消下载的资源(不能在下载过程中清除，被占用了，所以这里执行)
+                // isCanelend = true;
+                 //正在编辑资源和取消资源时返回
+                 if (Curstate == State.EDIT || Curstate ==State.UNDOWNlOAD || Curstate == State.CANCELUNZIP) return;
+                 if (Curstate == State.DOWNLOAD || Curstate == State.UNZIP)
+                 {
+                     if (ClickButton.name != name)
+                     return;
+                 } 
+                 ClickButton = newButton;
+                 //取消下载
+                 //如果正在下载,取消下载
+                 if (Curstate == State.DOWNLOAD)
+                 {
+                     SDK.CancelLoad("model/" + name);
+                     GameObject mask = newButton.transform.FindChild("Mask").gameObject;
+                     newButton.transform.FindChild("Tip").gameObject.SetActive(true);
+                     mask.SetActive(true);
+                     mask.GetComponent<Image>().fillAmount = 1;
+                     cancelList.Add(name);
+                     Curstate = State.UNDOWNlOAD;
+                     return;
+                 }
+                 //如果正在解压，取消解压
+                 else if (Curstate == State.UNZIP)
+                 {
+                     GameObject mask = newButton.transform.FindChild("Mask").gameObject;
+                     newButton.transform.FindChild("Tip").gameObject.SetActive(true);
+                     mask.SetActive(true);
+                     mask.GetComponent<Image>().fillAmount = 1;
+                     Curstate = State.CANCELUNZIP;
+                     cancelList.Add(name);
+                     return;
+                 }
+                 else
+                     //检查下载资源
+                     CheckResoucre(name);
              };
             //长按按钮
             newButton.GetComponent<DownTime>().DownEvent = (end) =>
@@ -274,7 +325,7 @@ public class Lobby : MonoBehaviour
                 }
                 Curstate = State.EDIT;
             };
-            //记录选中场景
+            //记录编辑选中场景
             newButton.transform.FindChild("Select").GetComponent<Toggle>().onValueChanged.AddListener((bool sel) =>
             {
                 if (sel == true & !delReList.Contains(newButton.name))
@@ -296,6 +347,38 @@ public class Lobby : MonoBehaviour
             coverobj.name = name;
         }
     }
+    // 返回取消下载资源
+    private void OnCancelLoad(string name)
+    {
+        Curstate = State.NOMAL;
+        isCanelend = true;
+    }
+    //清除取消的下载资源
+    void ClearCanelRe()
+    {
+        if (isCanelend == false) return;
+        try
+        {
+            while (cancelList.Count > 0)
+            {
+                if (File.Exists(SimpleFramework.Util.DataPath + cancelList[0] + ".upk"))
+                    File.Delete(SimpleFramework.Util.DataPath + cancelList[0] + ".upk");
+                if (Directory.Exists(SimpleFramework.Util.DataPath + cancelList[0]))
+                    Directory.Delete(SimpleFramework.Util.DataPath + cancelList[0], true);
+                if (Directory.Exists(SimpleFramework.Util.DataPath + "lua/" + cancelList[0] + ".upk") && SimpleFramework.AppConst.TestMode == false)
+                    Directory.Delete(SimpleFramework.Util.DataPath + "lua/" + cancelList[0] + ".upk", true);
+                if (Directory.Exists(SimpleFramework.Util.DataPath + "lua/" + cancelList[0]) && SimpleFramework.AppConst.TestMode == false)
+                    Directory.Delete(SimpleFramework.Util.DataPath + "lua/" + cancelList[0], true);
+                cancelList.RemoveAt(0);
+                ClickButton.transform.FindChild("Tip").gameObject.SetActive(false);
+                isCanelend = false;
+
+            }
+        }
+        catch
+        {
+        }
+    }
     //检查场景资源
     void CheckResoucre(string name)
     {
@@ -314,16 +397,25 @@ public class Lobby : MonoBehaviour
         //如果没有资源
         else
         {
-            if (Directory.Exists(SimpleFramework.Util.DataPath + "lua/" + SimpleFramework.AppConst.CurSceneName))
+            if (Directory.Exists(SimpleFramework.Util.DataPath + "lua/" + SimpleFramework.AppConst.CurSceneName)&&SimpleFramework.AppConst.TestMode==false)
                 Directory.Delete(SimpleFramework.Util.DataPath + "lua/" + SimpleFramework.AppConst.CurSceneName, true);
             if (Directory.Exists(SimpleFramework.Util.DataPath + SimpleFramework.AppConst.CurSceneName))
                 Directory.Delete(SimpleFramework.Util.DataPath + SimpleFramework.AppConst.CurSceneName, true);
             //下载lua压缩包
-            SDK.DownLoadToLocal("lua/" + SimpleFramework.AppConst.CurSceneName + ".upk", SimpleFramework.Util.DataPath + "lua", OnReleaseLua, false);
+            if (SimpleFramework.AppConst.TestMode == false)
+            {
+                SDK.DownLoadToLocal("lua/" + SimpleFramework.AppConst.CurSceneName + ".upk", SimpleFramework.AppConst.AppName + "/lua", OnReleaseLua, false);
+            }
+            Curstate = State.DOWNLOAD;
+            isMaskAni = true;
+            maskAniPos = 1;
             //下载资源压缩包
-            SDK.DownLoadToLocal("model/" + SimpleFramework.AppConst.CurSceneName + ".upk", SimpleFramework.Util.DataPath, OnReleaseResoucre);
+            SDK.DownLoadToLocal("model/" + SimpleFramework.AppConst.CurSceneName + ".upk", SimpleFramework.AppConst.AppName, OnReleaseResoucre);
+           
+
         }
     }
+
     //解压lua压缩包
     void OnReleaseLua(bool end)
     {
@@ -332,20 +424,60 @@ public class Lobby : MonoBehaviour
     //解压资源压缩包
     void OnReleaseResoucre(bool end)
     {
+        Image mask = ClickButton.transform.FindChild("Mask").GetComponent<Image>();
+        mask.fillAmount = 0.5f;
         UPKFolder.UnPackFolderAsync(SimpleFramework.Util.DataPath + SimpleFramework.AppConst.CurSceneName + ".upk", SimpleFramework.Util.DataPath, ResProgress);
     }
     //资源解压进程
     void ResProgress(long all, long now)
     {
+        
+       
         double progress = (double)now / all;
-       // Debug.Log("资源释放进度为: " + progress);
-        GUIDebug.info = "资源释放进度为: " + progress;
+        Debuger.Log("资源释放进度为: " + progress);
         if (progress >= 1)
         {
-            loadreend = true;
-            GameObject Group = transform.FindChild("Mask/Content/BookGroup").gameObject;
-            Group.transform.FindChild("Mask").gameObject.SetActive(true);
+            isCanelend = true;
+            Curstate = State.NOMAL;
+            return;
         }
+        if (Curstate == State.CANCELUNZIP)
+        {
+            releaseReVal = 1;
+            return;
+        }
+        Curstate = State.UNZIP;
+        isMaskAni = true;
+        maskAniPos = 2;
+        releaseReVal = 0.5f - (float)progress / 2;
+
+
+
+    }
+    void MaskAni()
+    {
+        if (isMaskAni == false) return;
+        Image  mask = ClickButton.transform.FindChild("Mask").GetComponent<Image>();
+        if (maskAniPos == 1)
+        {
+            if( mask.fillAmount > 0.5f)
+                mask.fillAmount -= 0.01f;
+            else isMaskAni = false;
+        } 
+        else if (maskAniPos == 2 )
+        {
+            mask.fillAmount = releaseReVal;
+            if (mask.fillAmount <= 0)
+            {
+                mask.gameObject.SetActive(false);
+                mask.fillAmount = 1;
+                isMaskAni = false;
+                Curstate = State.NOMAL;
+            }
+        }
+
+           
+
     }
     //lua解压进程
     void LuaProgress(long all, long now)
